Add momnet.js as a dependency
The moment library provides an easy way of working with dates.

I noticed that this project uses the `^` version range operator for its
dependencies. I opted to use strict version numbers to reduce the chance
of pulling down compromised/broken dependencies so that you get
reproducible builds locally and in CI.