Add ESLint as a dev dependency with a modified version of the recommended vue2.x settings + fix linter issues.
The project was lacking a good frontend linter which is something I think all serious projects should have today.

This commit adds ESLint as a dev dependency with a modified version of the Vue 2.x recommended settings.Additionally, it runs the --fix and sorts out some issues. I decided to modify the settings to accomodate for some of the previous design decisions made and to not introduce too many changes at once. The settings could perhaps be removed one by one to slowly migrate the project to the recommended Vue settings.