Add Larastan as a dev dependency with basic settings to get strated + fixed linter issues.
Just like with ESLint, the project was lacking a tool for static analysis in php, which I think all projects should have.

This commit adds Larastan as a dev dependency with some basic settings. Additionally, it contains small fixes to various files to make the linter happy. If the linter is happy, I'm happy.