Fix bug in ClientFactory, raised by seeders and caused by a typo.
The ClientFactory contained a typo in the address field.
The type was also present in the ClientsController.
However, not in the database which makes this less of an issue to fix.

This commit fixes the issue by correcting the typo from adress -> address in the ClientFactory and the ClientsController.
I'm not sure what has coused this typo, but since the database column is spelled correctly, this change shoudl be engouraged.
I did not find any other occurances of the typo.