Fix bug where client bookings did not show up + added tests to catch the bug.
Previously the ClientShow component expected the client data to include the bookings, while the controller method did not query any client bookings.

This commit fixes the issue by utilizing the `bookings` relation in the client query, used in the show method.
It also introduces a test to catch the same bug reappearing in the future. The commit also includes the "tests" directory in the static analysis.