Add database migration to create a foreign key on the clients table, to the users table.
This change allows a client to be assigned to a specific user, i.e, a way to separate client between users. The foreign key is set to "null on delete" as if a user is deleted, we probably don't want the clients to de deleted as well. Instead we want to be able to reassign them to other users. In addition to the poreign key contraint, the new field is also indexed to enchance query performance. The commit also adds the "database" directory to the static analysis.

Note: All the seeded clients will currently be assinged to the first user.

client user id