Add order to the clients returned by the ClientsController + Tests.
Previously the index method of the ClientsController returned the clients in a somewhat random order, which is obviously not what we or the user wants.

This commit orders the clients, starting from latest. The commit also includes a test to assert that the latest client is always first in the data returned.