Add client form validation, both client side and server side.
Previously the user was able to post any data in the new client form. This was a problem as we started noticing users entering random data.

This commit fixes the issue by introducing client-side and server-side validation of the new client data. The validation rules enforced by the feature are:
1. The name is required and can have a maxiumu length of 190 character.
2. The email address must be an valid email address.
3. The phone number must be a valid phone number including only digits, spaces and a plus sign.
4. Either the email address or the phone number is required.

swuash to validation

asd